Wrap the frequency and community views in a query to know whether or not the currentUser is allowed to post.
I'm having to refactor some of the view logic because we need to know certain information about the currentUser's relationship to the view before we render anything. For example, if the currentUser isn't a member of a community, they shouldn't see a story composer.

So I've wrapped the community and frequency views in a higher level query that gets this data first and determines what to render. This still needs more refeactoring and we're moving back towards a world of only needing a single query per view, rather than multiple queries for each component. Tbd.