Make session cookie expire after one year
People are being signed out after a day no matter what they do, which is
obviously suboptimal.

This changes the behaviour, now the cookie only expires after one year,
and the expire date is refreshed whenever a new request comes in from
that user.

Essentially, it now expires when a user is completely inactive for a year.