Fix development environment
The issue was that we're now importing the client-side code on the
server for SSR, so when you change anything on the frontend the server
and the client would restart at the same time. The client was a bit
faster than the server with rebuilding so the browser would refresh with the new code
and fetch data but the server was still starting so you got those
annoying errors.

This fixes it by ignoring changes in `src/` from the server if we're not
testing SSR (i.e. if `process.env.SSR` isn't set) and we're not in
production.

Closes #1467, sorry about all the troubles!

Ref #1449 for good measure.