Defer Raven.js <script>
The `defer` attribute makes it so that a script is only executed _after_
the HTML has been parsed and rendered. Contrary to the `async`
attribute, `defer` also guarantees that the ordering of script
executions is consistent. (i.e. a later script in the HTML will execute
after the deferred one)

Looking at some [WebPageTest results](https://goo.gl/LACohE)
(specifically the [filmstrim
view](https://www.webpagetest.org/video/compare.php?tests=171012_PZ_92a736797d24b4b5861d897ca52697d4-r:1-c:0)) one can clearly see that the server-side rendered HTML is only rendered to the screen after `raven.js` is loaded and executed.

That's super stupid, as we don't want Raven.js to block the initial HTML
render at all! We want it to be loaded before the main bundle so any
errors in the initialization code are reported to Sentry, but it should
not block the initial HTML render.

That's why the `defer` attribute on the `script` tag makes sense here:

1. It doesn't block the initial render anymore
2. It still guarantees that our `bundle.js` is executed after `raven.js`

That's perfect for our needs and should make the initial pageload much
faster.