Add unhandled rejection listening to hyperion
This should also fix the reporting to Sentry @brianlovin, the issue was
that Raven.captureException doesn't return a Promise so the `await`
didn't work.

I worked around that by wrapping it in a Promise that passes resolve()
as the callback to `Raven.captureException`, which is called once the
exception is reported.