Clean up flow errors
The reason that the variable assignment fixes those flow errors is
because flow doesn't know whether the `searchIndex.search({ filters })`
function will mutate the filters, so it just assumes it deletes it.

By asssigning the part we actually use to a variable, even if the search
would go `delete filters.whatever` we'd still have the value as a string
in the variable and can thusly safely use it.