Update PR template
The checklist-style for deployment made the pr overview screen really
useless because every PR now had a checklist associated with it.

Instead, let's just delete the servers that don't need deployment from
the list. Same thing for the db migrations, if people need to run one
after merging leave the section in, otherwise delete it.