Fix duplication of local messages
This patch fixes the local message duplication by checking for message
content equality. While that's pretty brittle (if we change anything
about the content of a message on the server, anything at all, it'll
break again), it works for now until we figure out a better way to fix
this.