Add and implement sendPushNotifications queue
@brianlovin note how I avoid adding duplicate jobs in
athena/listeners/notifications.js: By assigning a custom `jobId` that's
specific to the usersNotificationsJoin. Even if multiple instances of
athena are running and listening to new notifications, if a notification
comes in the job is keyed by its ID, so that if two instance add the
same job it's not duplicated.

Food for though, maybe that'll help with your current work in Pluto
potentially.