Run desktop CD from the alpha branch
As @brianlovin noticed in #3199, the CD for the desktop app wasn't
working because CircleCI only runs builds for commits that are
associated with a PR.

While we could set CircleCI to run builds for all branches, that's a
very suboptimal solution as it would run a ton of unnecessary builds.

I came to wondering, why does the CD for alpha work if it doesn't build
branches? Hidden in the CircleCI docs I found this snippet:

> Note: For your default branch, we will always build all commits.

Which explains why alpha CD works, but production CD doesn't.

This patch is a very simple solution to this problem: we run desktop CD
on the alpha branch. That is fine because the desktop CD only creates a
draft GitHub release and we have to manually publish that release for
anybody to get the update, so it doesn't matter if we CD from alpha or
production!