Don't server-side render for desktop app
The desktop app requests the same paths as the browser would, which
means previously it would also get the server-side rendered HTML.

The problem is that the server-side rendered frontend doesn't know that
it's being rendered for the desktop app, and thusly renders the standard
frontend, which looks crappy in the desktop app.

This patch disables server-side rendering enterily when a req from the
desktop app comes in, which fixes that issue.