Fix forceScrollToBottom logic
We only want to force scroll to bottom on short threads. Now that we
have message pagination, if a thread has > 50 messages we automatically
show the newest message at the top—so we shouldn't scroll the thread to
the bottom, no matter if you're the author, a participant or if you've
seen it before.