Replace all props.theme function interpolations with static interpolations
styled-components has a fast path built-in: when styles don't have any
function interpolations, we never recompute them since they can't change.
This is especially relevant when re-rendering the same components over
and over and over, e.g. when switching between threads.

This patch was done automatically, and moves all of our simple function
interpolations in the form of ${props => props.theme.x} to static
interpolations that just do ${theme.x}. This allows us to hit the fast
path in many more cases, making the app re-render faster. This is most
noticable in the dashboard when switching threads.